// HFY.MvcWebUI.Angular/ClientApp/src/app/lazy-general-page/classes/lastik-olcumler/models/lastik-olcum-ekle.ts
export class LastikOlcumEkle {
    LastikID: number;
    AracID:number;
    Tarih: Date | string;
    AracKilometre: number;
    LastikKilometre: number;
    DisDerinligiJSON: string;
    GuvenliDisSeviyesi: number;
    Basinc: number;
    BasincAlinamadi: boolean;
    LastikKonumID:number;
    LastikPozisyonID:number;
    TavsiyeBasinc:number;
    GozlemJSON: string;
    GozlemYapildiMi:boolean;
}